// frontend/src/app/models/reportInputData.ts
export class ReportData{
    date_of_report: Date;
    doctors_name: string;
    specializzazione: string;
    reason_for_comming: string;
    diagnosis: string;
    therapy: string;
    next_session: Date;
    patient_id: string;
    date_of_schedule: Date;
    doctor_id: Object;
    date1:string;
    date2: string
}