// frontend/src/app/models/myapoint.ts
export class MyApointmentsdata{
    _id: Object;
    atient_id: string;
    servics: string;
    date_start: Date;
    date_end: Date;
    firstname: string;
    lastname: string;
    status: string;
    doctorts_note: boolean;
    cancled: boolean;
    cancled_note: string;
    doctor_firstname: string;
    doctor_lastname: string;
    office_branch: string;

}